Ingreso de nuevos ajustes

+++ Ejercicio 1.py
@@ -1,6 +1,6 @@
-Nota1 = int(input("Ingrese la primera nota: "))
-Nota2 = int(input("Ingrese la segunda nota: "))
-Nota3 = int(input("Ingrese la tercera nota: "))
-Nota4 = int(input("Ingrese la cuarta nota: "))
-Nota5 = int(input("Ingrese la quinta nota: "))
+Nota1 = float(input("Ingrese la primera nota: "))
+Nota2 = float(input("Ingrese la segunda nota: "))
+Nota3 = float(input("Ingrese la tercera nota: "))
+Nota4 = float(input("Ingrese la cuarta nota: "))
+Nota5 = float(input("Ingrese la quinta nota: "))
